+++ pgjdbc/src/main/java/org/postgresql/core/PGStream.java
@@ -21,6 +21,7 @@ import java.io.OutputStream;
 import java.io.Writer;
 import java.net.InetSocketAddress;
 import java.net.Socket;
+import java.net.SocketTimeoutException;
 import java.sql.SQLException;
 import javax.net.SocketFactory;
 
@@ -42,6 +43,11 @@ public class PGStream implements Closeable, Flushable {
   private VisibleBufferedInputStream pg_input;
   private OutputStream pg_output;
   private byte[] streamBuffer;
+
+  private long nextStreamAvailableCheckTime;
+  // This is a workaround for SSL sockets: sslInputStream.available() might return 0
+  // so we perform "1ms reads" once in a while
+  private int minStreamAvailableCheckDelay = 1000;
 
   private Encoding encoding;
   private Writer encodingWriter;
@@ -109,7 +115,29 @@ public class PGStream implements Closeable, Flushable {
    * @throws IOException if something wrong happens
    */
   public boolean hasMessagePending() throws IOException {
-    return pg_input.available() > 0 || connection.getInputStream().available() > 0;
+    if (pg_input.available() > 0) {
+      return true;
+    }
+    // In certain cases, available returns 0, yet there are bytes
+    long now = System.currentTimeMillis();
+    if (now < nextStreamAvailableCheckTime && minStreamAvailableCheckDelay != 0) {
+      // Do not use ".peek" too often
+      return false;
+    }
+    nextStreamAvailableCheckTime = now + minStreamAvailableCheckDelay;
+    int soTimeout = getNetworkTimeout();
+    setNetworkTimeout(1);
+    try {
+      return pg_input.peek() != -1;
+    } catch (SocketTimeoutException e) {
+      return false;
+    } finally {
+      setNetworkTimeout(soTimeout);
+    }
+  }
+
+  public void setMinStreamAvailableCheckDelay(int delay) {
+    this.minStreamAvailableCheckDelay = delay;
   }
 
   /**