+++ pgjdbc/src/test/java/org/postgresql/test/jdbc2/DatabaseEncodingTest.java
@@ -260,7 +260,7 @@ public class DatabaseEncodingTest {
         // Seven-byte illegal sequences
         {(byte) 0xfe}, // Can't have a seven-byte sequence.
 
-        // Eigth-byte illegal sequences
+        // Eighth-byte illegal sequences
         {(byte) 0xff}, // Can't have an eight-byte sequence.
     };